+++ test/value_test.py
@@ -4,44 +4,50 @@
 #
 
 from dataclasses import dataclass
-from fxn import Dtype, Value
+from fxn import Function, Dtype, Value
 from numpy.random import randn
-from pathlib import Path
 from pydantic import BaseModel
 
 def test_serialize_null ():
-    value = Value.from_value(None, "null")
+    fxn = Function()
+    value = fxn.predictions.from_value(None, "null")
     assert value.type == Dtype.null
     assert value.data == None
     assert value.shape == None
 
 def test_serialize_value ():
+    fxn = Function()
     value = Value(data="https://fxn.ai", type=Dtype.int64)
-    serialized = Value.from_value(value, "value")
+    serialized = fxn.predictions.from_value(value, "value")
     assert serialized == value
 
 def test_serialize_tensor ():
+    fxn = Function()
     tensor = randn(1, 3, 720, 1280)
-    value = Value.from_value(tensor, "tensor")
+    value = fxn.predictions.from_value(tensor, "tensor")
     assert value.type == str(tensor.dtype)
     assert value.shape == list(tensor.shape)
 
 def test_serialize_string ():
-    value = Value.from_value("Hello fxn!", "string")
+    fxn = Function()
+    value = fxn.predictions.from_value("Hello fxn!", "string")
     assert value.type == Dtype.string
 
 def test_serialize_float ():
-    value = Value.from_value(3.1415, "integer")
+    fxn = Function()
+    value = fxn.predictions.from_value(3.1415, "integer")
     assert value.type == Dtype.float32
     assert value.shape == []
 
 def test_serialize_int ():
-    value = Value.from_value(24, "integer")
+    fxn = Function()
+    value = fxn.predictions.from_value(24, "integer")
     assert value.type == Dtype.int32
     assert value.shape == []
 
 def test_serialize_bool ():
-    value = Value.from_value(True, "boolean")
+    fxn = Function()
+    value = fxn.predictions.from_value(True, "boolean")
     assert value.type == Dtype.bool
     assert value.shape == []
 
@@ -55,11 +61,13 @@ def test_serialize_list ():
         state: str
     person = Person(name="Jake", age=32)
     city = City(name="New York", state="New York")
-    value = Value.from_value(["cat", 365, person, city], "list")
+    fxn = Function()
+    value = fxn.predictions.from_value(["cat", 365, person, city], "list")
     assert value.type == Dtype.list
 
 def test_serialize_dict ():
-    value = Value.from_value({ "language": "typescript" }, "dict")
+    fxn = Function()
+    value = fxn.predictions.from_value({ "language": "typescript" }, "dict")
     assert value.type == Dtype.dict
 
 def test_serialize_model ():
@@ -67,5 +75,6 @@ def test_serialize_model ():
         wheels: int
         mileage: float
     vehicle = Vehicle(wheels=4, mileage=324.2)
-    value = Value.from_value(vehicle, "pydantic")
+    fxn = Function()
+    value = fxn.predictions.from_value(vehicle, "pydantic")
     assert value.type == Dtype.dict